+++ apps/modorix-backend/src/domain/block-reason.service.ts
@@ -1,6 +1,6 @@
 import { BlockReason } from '@modorix-commons/models/block-reason';
 import { Injectable } from '@nestjs/common';
-import { BlockReasonsRepository } from '../infrastructure/block-reason.repository';
+import { BlockReasonsRepository } from '../infrastructure/repositories/block-reason.repository';
 
 @Injectable()
 export class BlockReasonsService {
