+++ apps/modorix-backend/src/domain/block-x-user.service.ts
@@ -1,8 +1,8 @@
 import { BlockXUserRequest, XUser } from '@modorix-commons/models/x-user';
 import { Injectable } from '@nestjs/common';
-import { BlockReasonsRepository } from '../infrastructure/block-reason.repository';
-import { BlockXUsersRepository } from '../infrastructure/block-x-user.repository';
-import { GroupsRepository } from '../infrastructure/groups.repository';
+import { BlockReasonsRepository } from '../infrastructure/repositories/block-reason.repository';
+import { BlockXUsersRepository } from '../infrastructure/repositories/block-x-user.repository';
+import { GroupsRepository } from '../infrastructure/repositories/groups.repository';
 import { BlockReasonError } from './errors/block-reason-error';
 import { XUserNotFoundError } from './errors/x-user-not-found-error';
 import { XUserNotInQueueError } from './errors/x-user-not-in-queue';
